+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @duxweb/dvha-core
 
+## 1.2.0
+
+### Minor Changes
+
+- feat: update vite8
+
+## 1.1.31
+
 ## 1.1.30
 
 ## 1.1.29