+++ packages/pro/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @duxweb/dvha-pro
 
+## 1.2.0
+
+### Minor Changes
+
+- feat: update vite8
+
+### Patch Changes
+
+- Updated dependencies
+  - @duxweb/dvha-naiveui@1.2.0
+  - @duxweb/dvha-core@1.2.0
+
+## 1.1.31
+
+### Patch Changes
+
+- feat: 更新到 vite8
+  - @duxweb/dvha-core@1.1.31
+  - @duxweb/dvha-naiveui@1.1.31
+
 ## 1.1.30
 
 ### Patch Changes