+++ packages/template/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @duxweb/dvha-template
 
+## 1.2.0
+
+### Minor Changes
+
+- feat: update vite8
+
+## 1.1.31
+
+### Patch Changes
+
+- feat: 更新到 vite8
+
 ## 1.1.30
 
 ## 1.1.29